backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
yfinance==1.3.0
pandas==2.2.1
numpy==1.26.4
anthropic>=0.21.0
python-dotenv==1.0.1
httpx==0.27.0
lxml==5.1.0
